Dockerized_Entry_Local/requirements.txt
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-phonenumber-field==3.0.1
gunicorn==20.0.3
idna==2.8
phonenumbers==8.10.23
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
twilio==6.33.1
urllib3==1.25.7
